requirements_cpu.txt
pandas==2.2.2
Flask==3.0.3
mysql-connector-python==8.3.0
Sympy==1.12
openpyxl==3.1.2
opencv-python==4.9.0.80
PyMuPDF==1.24.2
Pillow==10.3.0
spacy==3.7.4
torch==2.3.0
torchvision==0.18.0
easyocr==1.7.1
fuzzywuzzy==0.18.0
pyspellchecker==0.8.1
spacy-transformers==1.3.5
datetime==5.5
python-dotenv==1.0.1
passlib==1.7.4
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pytesseract==0.3.10
boto3==1.34.99
celery==5.4.0
redis==3.5.3
gunicorn==22.0.0
